правки багів

+++ SharedLib/WDB_SQLite.cs
@@ -201,7 +201,7 @@ namespace SharedLib
                             RW.Price = MPI.GetPrice(Res.Price, Res.IsIgnoreMinPrice == 0, Res.CodePs > 0);
                             RW.TypePrice = MPI.typePrice;
                             RW.ParPrice1 = Res.CodePs;
-                            RW.ParPrice2 = (int)Res.TypeDiscont;
+                            RW.ParPrice2 = (long)Res.TypeDiscont;
                             RW.ParPrice3 = Res.Data;
                             RW.Priority = Res.Priority;
                         }
